Separate String functions into own file
